Created new file for utility functions
